requirements.txt
torch>=1.12.0
torchvision>=0.13.0
timm>=0.6.0
numpy>=1.21.0
pillow>=8.0.0
pyyaml>=5.4.0
tqdm>=4.62.0
tensorboard>=2.9.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
pytest>=6.2.0
black>=21.0
flake8>=3.9.0
jupyter>=1.0.0
ipykernel>=6.0.0
